requirements.txt
kucoin-python
python-dotenv